Imlemented the fetch stage and decode stage. Also created a Send Data event. Cretaed all 64 registers. Finished the flow between the stages. Also started to convert the asm instructions to binary.

+++ src/memory.hh
@@ -36,9 +36,8 @@ public:
 // Class for each Instruction in the instruction memory
 class Instruction{
 protected:
-    uint8_t binary; // Holds the 8 bit instruction for the memory loaction. 4 memory locations equal 32 bits.
+    uint32_t binary; // Holds the 8 bit instruction for the memory loaction. 4 memory locations equal 32 bits.
     std::string instructionType;
-    // Register *operand; // May not need this
 
 public:
     Instruction(uint32_t b) : binary(b) {}
